Renamed sim.py
