+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/untitled3.iml" filepath="$PROJECT_DIR$/.idea/untitled3.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/rain_heightmap.iml" filepath="$PROJECT_DIR$/.idea/rain_heightmap.iml" />
     </modules>
   </component>
 </project>
